refactor: seperate files for const tstep observer and rename struct
